+++ src/Form/BackupForm.php
@@ -19,20 +19,25 @@
  * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
  * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ * @see         MeCms\Controller\Admin\BackupsController::add()
  */
 namespace MeCms\Form;
 
 use Cake\Form\Form;
 use Cake\Network\Exception\InternalErrorException;
+use MeCms\Model\Validation\AppValidator;
 use MysqlBackup\Utility\BackupExport;
 
 /**
- * BackupForm class.
- *
- * It is used by `MeCms\Controller\Admin\BackupsController::add()`.
+ * BackupForm class
  */
 class BackupForm extends Form
 {
+    /**
+     * @var \MysqlBackup\Utility\BackupExport
+     */
+    protected $_BackupExport;
+
     /**
      * Defines the validator using the methods on Cake\Validation\Validator or
      * loads a pre-defined validator from a concrete class.
@@ -41,45 +46,47 @@ class BackupForm extends Form
      */
     protected function _buildValidator(\Cake\Validation\Validator $validator)
     {
-        $validator = new \MeCms\Model\Validation\AppValidator();
+        $validator = new AppValidator;
 
         //Filename
-        $validator->requirePresence('filename')
-            ->remove('filename', 'validateUnique')
-            ->add('filename', 'validExtension', [
-                'rule' => function ($value, $context) {
-                    $extensions = array_map(function ($v) {
-                        return preg_quote($v, '/');
-                    }, ['sql', 'sql.gz', 'sql.bz2']);
-
-                    return (bool)preg_match(sprintf('/\.(%s)$/i', implode('|', $extensions)), $value);
-                },
-                'message' => __d('me_cms', 'Valid extensions: {0}', 'sql, sql.gz, sql.bz2'),
-            ])
-            ->add('filename', [
-                'lengthBetween' => [
-                    'message' => __d('me_cms', 'Must be between {0} and {1} chars', 3, 100),
-                    'rule' => ['lengthBetween', 3, 100],
-                ],
-            ]);
+        $validator->add('filename', 'extension', [
+            'message' => __d('me_cms', 'Valid extensions: {0}', 'sql, sql.gz, sql.bz2'),
+            'rule' => function ($value) {
+                return (bool)preg_match('/[^\.]\.sql(\.(bz2|gz))?$/', $value);
+            },
+        ])->requirePresence('filename');
 
         return $validator;
     }
 
     /**
+     * Gets a `BackupExport` instance
+     * @return \MysqlBackup\Utility\BackupExport
+     * @uses $_BackupExport
+     */
+    protected function _getBackupExportInstance()
+    {
+        if (empty($this->_BackupExport)) {
+            $this->_BackupExport = new BackupExport;
+        }
+
+        return $this->_BackupExport;
+    }
+
+    /**
      * Used by `execute()` to execute the form's action
      * @param array $data Form data
-     * @return bool
+     * @return string|bool Filename or `false` on failure
      * @uses MysqlBackup\Utility\BackupExport::filename()
      * @uses MysqlBackup\Utility\BackupExport::export()
+     * @uses _getBackupExportInstance()
      */
     protected function _execute(array $data)
     {
         try {
-            $backup = new BackupExport();
-            $backup->filename($data['filename']);
-
-            return $backup->export();
+            return $this->_getBackupExportInstance()
+                ->filename($data['filename'])
+                ->export();
         } catch (InternalErrorException $e) {
             return false;
         }