+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.15 branch
 ### 2.15.3
+* fixed code for `BackupForm` class. Added `_getBackupExportInstance()` method;
 * fixed code and a little bug for `ContactForm` and `ContactFormMailer` classes;
 * added tests for `ContactForm`, `ContactFormMailer` and `Mailer` classes.